requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-decouple==2.1
django-phonenumber-field==8.1.0
django-rest-knox==5.0.2
djangorestframework==3.16.0
geographiclib==2.0
geopy==2.4.1
joblib==1.5.0
numpy==2.2.5
pandas==2.2.3
phonenumbers==9.0.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
